+++ packages/azure/src/__tests__/AzureAccount.test.ts
@@ -222,7 +222,7 @@ describe('Azure Account', () => {
         delete AZURE.SUBSCRIPTION_CHUNKS
       })
 
-      it('fetches data only for specific subscriptions when specified subscriptions are set', async () => {
+      it('fetches data only for specific subscriptions when specified environment variable subscriptions are set', async () => {
         const mockCredentials = {
           clientId: 'test-client-id',
           secret: 'test-client-secret',
@@ -294,6 +294,76 @@ describe('Azure Account', () => {
 
         // Reset the config
         delete AZURE.SUBSCRIPTIONS
+      })
+
+      it('fetches data only for specific subscriptions when specified parameter subscriptions are set', async () => {
+        const mockCredentials = {
+          clientId: 'test-client-id',
+          secret: 'test-client-secret',
+          domain: 'test-tenant-id',
+        }
+        ;(createCredentialsSpy as jest.Mock).mockResolvedValue(mockCredentials)
+
+        const subscriptions = ['sub-1', 'sub-2', 'sub-3']
+
+        subscriptions.forEach((subscriptionId) => {
+          mockListSubscriptions.get.mockReturnValueOnce({ subscriptionId })
+        })
+
+        const mockEstimates: EstimationResult[] = [
+          {
+            timestamp: startDate,
+            serviceEstimates: [
+              {
+                kilowattHours: 0.09313874999999999,
+                co2e: 0.000021235635,
+                usesAverageCPUConstant: true,
+                cloudProvider: 'AZURE',
+                accountId: testAccountId,
+                accountName: testAccountName,
+                serviceName: 'Virtual Machines',
+                cost: 5,
+                region: 'UK South',
+              },
+            ],
+            periodStartDate: startDate,
+            periodEndDate: endDate,
+            groupBy: grouping,
+          },
+        ]
+
+        ;(getEstimatesSpy as jest.Mock).mockResolvedValue(mockEstimates)
+        const getDataForSubscriptionSpy = jest.spyOn(
+          AzureAccount.prototype,
+          'getDataForSubscription',
+        )
+
+        // when
+        const azureAccount = new AzureAccount()
+        await azureAccount.initializeAccount()
+        const results = await azureAccount.getDataFromConsumptionManagement(
+          startDate,
+          endDate,
+          grouping,
+          subscriptions,
+        )
+
+        // Results should be the same as the mockEstimates array repeated35 times
+        const expectedEstimates = Array.from(
+          { length: 3 },
+          () => mockEstimates[0],
+        )
+
+        expect(results).toEqual(expectedEstimates)
+        // getDataForSubscription should have been called 3 times (once for each given subscription ID)
+        subscriptions.forEach((subscriptionId) => {
+          expect(getDataForSubscriptionSpy).toHaveBeenCalledWith(
+            startDate,
+            endDate,
+            subscriptionId,
+            grouping,
+          )
+        })
       })
     })
   })